Fix mmap ENOMEM by replace "atoi" with "atol": When input argv[2]="2253980245" (0x86590255), atoi() cast it to "18446744071668564565" (0xffffffff86590255), so mmap ENOMEM

+++ usemem_remap.c
@@ -33,7 +33,7 @@ int main (int argc, char *argv[])
 		exit(1);
 	}
 
-	size = atoi(argv[2]);
+	size = atol(argv[2]);
 	no_of_pages = size / pagesize;
 
 	if ((fd = open(argv[1], O_RDWR | O_CREAT, 0666)) == -1) {
